Create README.md
The world of blockchain and cryptocurrencies has been characterized by relentless innovation and paradigm-shifting ideas. In this dynamic landscape, the HyperCycle ecosystem emerges as a groundbreaking and transformative concept, setting its course to reshape the very essence of digital assets. Comprising two fundamental tokens, Hyperlink (HLINK) and Hyperchian (HCHAIN), the HyperCycle ecosystem represents a compelling response to the ever-evolving needs and demands of the cryptocurrency sphere.